Full refactory for REST-oriented project.

+++ src/main/java/ru/evaproj/analyst/history/dto/CandleDto.java
@@ -8,10 +8,11 @@ public class CandleDto {
 
     private Long timestamp;
 
-    private Long open;
-    private Long close;
-    private Long high;
-    private Long low;
+    private Double open;
+    private Double close;
+    private Double high;
+    private Double low;
+
     private Long volume;
 
 }
